refactor: moved bike-model server actions to /bikeModels
Moved the servers actions from /actions/bikes to /actions/bikeModels for more clarity.
